Adding mutations
Add submit tweet and delete tweet mutations

Add simple styling
