+++ app-client/appsync-client/src/components/helpers.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+export const Div = styled.div`
+  margin-top: 30px;
+  margin-bottom: 30px;
+`;
+
+export const Container = styled.div`
+  width: 100%;
+  max-width: 100%;
+  text-align: left;
+  background-color: #ffffff;
+  border-radius: 2px;
+  box-shadow: 0px 1px 1px 0 rgba(0, 0, 0, 0.3);
+
+  p {
+    padding: 10px;
+  }
+`;
