+++ app-client/appsync-client/src/queries/index.js
@@ -1,0 +1,60 @@
+import gql from 'graphql-tag';
+
+export const UserQuery = gql`
+  query UserQuery(
+    $handle: String!
+    $consumer_key: String!
+    $consumer_secret: String!
+  ) {
+    getUserTwitterFeed(
+      handle: $handle
+      consumer_key: $consumer_key
+      consumer_secret: $consumer_secret
+    ) {
+      name
+      location
+      handle
+      favourites_count
+      description
+      followers_count
+      friends_count
+      followers
+      topTweet {
+        tweet_id
+        tweet
+        retweeted
+        retweet_count
+        favorited
+      }
+    }
+  }
+`;
+
+export const UserTweetsQuery = gql`
+  query UserQuery(
+    $handle: String!
+    $consumer_key: String!
+    $consumer_secret: String!
+  ) {
+    getUserTwitterFeed(
+      handle: $handle
+      consumer_key: $consumer_key
+      consumer_secret: $consumer_secret
+    ) {
+      tweets(limit: 10) {
+        items {
+          tweet
+          tweet_id
+          retweet_count
+          retweeted
+          favorited
+        }
+        nextToken
+      }
+    }
+  }
+`;
+
+export default {
+  UserQuery,
+};
